gitignore: CUSTOM-LOCAL.config; clean mycron file only, if present

+++ 8_set_cronjobs.sh
@@ -25,4 +25,4 @@ else
 fi
 
 # cleaning
-rm mycron
+[ -w mycron ] && rm mycron
